particle-analytics-quickstarts/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "particle-flat-observatory"
version = "0.1.0"
description = "Data pipeline accelerator for Particle Health flat data — schema generation, normalization, and analytics"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.21.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
observatory-generate-ddl = "observatory.generate_ddl:main"
particle-pipeline = "observatory.cli:app"

[project.optional-dependencies]
bigquery = [
    "google-cloud-bigquery>=3.40.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/observatory"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
